fix runtime object updates and worker imports
