+++ ui/src/lib/js-runner/index.ts
@@ -1,1 +1,2 @@
-export { WorkerNodeSystem, type WorkerMessage, type WorkerResponse } from './WorkerNodeSystem';
+export { WorkerNodeSystem } from './WorkerNodeSystem';
+export type { WorkerMessage, WorkerResponse } from './js-worker-types';
