+++ ui/src/objects/toggle/ToggleButtonNode.svelte
@@ -23,7 +23,9 @@
     };
   });
 
-  let isOn = $derived(data.value === true);
+  let isOn = $derived(
+    data.value === true || (data.value === undefined && data.params?.[0] === true)
+  );
 
   const borderColor = $derived(selected ? '!border-zinc-400' : '!border-zinc-600');
